add --interleave option for multi-agent plans
The agents can proceed on their own links as long as they do
not depend on anything done by the other agents. These points
are indicated in the linking plan as below:

(in links_for_agent_2_of_3.txt):
...
  14      1    3 Stonewall Jackson Statue -> 4 Congregation Beth Israel
  15      1    3 Stonewall Jackson Statue -> 2 Gen. Alexander Archer Vandegri
  16*     3    6 Catholic Church of the Holy Comforter -> 10 Merriweather Lewis Placard

  17* ____2   10 Merriweather Lewis Placard
              11 Albermarle County Courthouse

wait for: agent 1 to complete link 15 before proceeding
wait for: agent 3 to complete link 10 before proceeding
  18  ____2   10 Merriweather Lewis Placard
               1 Historic Jail House
...

Note that the completion order of some fields may change when
interleaving of actions is allowed. Thus the estimated link
and AP counts for the agents may be slightly off.